Adding a hell of a lot of css and database work
This is mostly the beginning of the process of the design of the
website, the database can be accessed now on pages thorugh EJS, as was
intended. This should be a jumping off point, just cause this is the
first commit for the official team. All big changes from here on out
should be done in separate branches with pull requests. So feel free to
edit this thing to make it amazing!